pyproject.toml
[tool.poetry]
name = "shorty"
version = "0.1.0"
description = ""
authors = ["Yegor <egorhs19@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask-RESTful = "^0.3.9"
pyshorteners = "^1.0.1"
pydantic = "^1.8.2"
flasgger = "^0.9.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-flask = "^1.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
